pyproject.toml
[tool.poetry]
name = "airflow-demo"
version = "0.1.0"
description = "Python project for Airflow demo"
authors = ["Leandro Szikora"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
apache-airflow = "2.10.2"
requests = "^2.32.3"
apache-airflow-providers-postgres = "5.10.0"
apache-airflow-providers-slack = "8.9.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
